requirements.txt
Flask==2.1.0
gunicorn==20.1.0
opentelemetry-exporter-gcp-trace==1.4.0
opentelemetry-propagator-gcp==1.4.0
opentelemetry-instrumentation-flask==0.34b0
opentelemetry-instrumentation-requests==0.34b0
google-cloud-trace==1.11.1
